pyproject.toml
[project]
name = "agent-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "tiktoken>=0.12.0",
]
